backend/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
aiofiles
python-multipart
pydantic
httpx
python-dotenv